+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.0 (Unreleased)
+
+* Adds Cloud Logging Service Endpoint to configuration
+* Fixes hide not working for Cloud Logging query
+* Fixes inability to retrieve projects displaying an error
+
 ## 1.0.0 (2023-01-17)
 
 Initial release.